@@ -13,37 +13,43 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "^5.1.1",
-    "@angular/cdk": "^5.0.1",
-    "@angular/common": "^5.1.1",
-    "@angular/compiler": "^5.1.1",
-    "@angular/core": "^5.1.1",
-    "@angular/forms": "^5.1.1",
-    "@angular/http": "^5.1.1",
-    "@angular/material": "^5.0.1",
-    "@angular/platform-browser": "^5.1.1",
-    "@angular/platform-browser-dynamic": "^5.1.1",
-    "@angular/router": "^5.1.1",
+    "@agm/snazzy-info-window": "^1.0.0-beta.2",
+    "@angular/animations": "^5.2.2",
+    "@angular/cdk": "^5.1.0",
+    "@angular/common": "^5.2.2",
+    "@angular/compiler": "^5.2.2",
+    "@angular/core": "^5.2.2",
+    "@angular/forms": "^5.2.2",
+    "@angular/http": "^5.2.2",
+    "@angular/material": "^5.1.0",
+    "@angular/platform-browser": "^5.2.2",
+    "@angular/platform-browser-dynamic": "^5.2.2",
+    "@angular/router": "^5.2.2",
+    "@ngx-loading-bar/http": "^1.1.0",
+    "@ngx-loading-bar/http-client": "^1.1.0",
     "angular2-csv": "^0.2.5",
     "core-js": "^2.5.3",
+    "file-saver": "^1.3.3",
     "hammerjs": "^2.0.8",
-    "rxjs": "^5.5.5",
-    "zone.js": "^0.8.14"
+    "rxjs": "^5.5.6",
+    "snazzy-info-window": "^1.1.1",
+    "xlsx": "^0.11.18",
+    "zone.js": "^0.8.20"
   },
   "devDependencies": {
     "@angular/cli": "1.5.5",
-    "@angular/compiler-cli": "^5.1.1",
-    "@angular/language-service": "^5.1.1",
+    "@angular/compiler-cli": "^5.2.2",
+    "@angular/language-service": "^5.2.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.94",
-    "codelyzer": "^4.0.2",
+    "@types/node": "^6.0.96",
+    "codelyzer": "^4.1.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.4.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
