@@ -1,32 +1,42 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, HostListener} from '@angular/core';
 import {RetrieveCompaniesService} from './retrieve-companies.service';
-import {CompanyInterface} from './company-interface';
+import {ExportService} from './export.service';
+import {QueryBuilderService} from './query-builder.service';
 
 @Component({
     selector: 'app-root',
     templateUrl: './app.component.html',
-    styleUrls: ['./app.component.css'],
-    providers: [RetrieveCompaniesService],
+    styleUrls: ['./app.component.scss'],
+    providers: [RetrieveCompaniesService, ExportService, QueryBuilderService],
 })
-export class AppComponent implements OnInit {
-    companies: CompanyInterface[];
-    params = [];
+export class AppComponent {
+    toggle = true;
+    screenWidth: number;
 
-    constructor(private retrieveCompaniesService: RetrieveCompaniesService) {
+    constructor() {
+        // set screenWidth on page load
+        this.screenWidth = window.innerWidth;
+        window.onresize = () => {
+            // set screenWidth on screen size change
+            this.screenWidth = window.innerWidth;
+        };
+
+        if (window.innerWidth <= 991) {
+            this.toggle = false;
+        }
     }
 
-    getCompanies(): void {
-        this.retrieveCompaniesService.getCompanies(this.params).subscribe(data => {
-            this.companies = data.companies as CompanyInterface[];
-        });
+    @HostListener('window:resize', ['$event'])
+    onResize(event) {
+        // medium devices
+        if (event.target.innerWidth <= 991 && this.toggle === true) {
+            this.toggle = false;
+        } else if (event.target.innerWidth > 991 && this.toggle === false) {
+            this.toggle = true;
+        }
     }
 
-    ngOnInit() {
-        this.getCompanies();
-    }
-
-    onFilter(event) {
-        this.params[event.name] = event.filter;
-        this.getCompanies();
+    toggleSidebar() {
+        this.toggle = !this.toggle;
     }
 }
