@@ -7,8 +7,12 @@ import {HttpClientModule} from '@angular/common/http';
 import {ListCompaniesComponent} from './list-companies/list-companies.component';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {SidebarComponent} from './sidebar/sidebar.component';
-import {AppRoutingModule} from './/app-routing.module';
+import {AppRoutingModule} from './app-routing.module';
 import {AgmCoreModule} from '@agm/core';
+import { AgmSnazzyInfoWindowModule } from '@agm/snazzy-info-window';
+import * as myGlobals from './globals';
+import { LoadingBarHttpClientModule } from '@ngx-loading-bar/http-client';
+import { LoadingBarModule } from '@ngx-loading-bar/core';
 import {
     MatAutocompleteModule,
     MatButtonModule,
@@ -44,6 +48,7 @@ import {
 } from '@angular/material';
 import {MapCompaniesComponent} from './map-companies/map-companies.component';
 import {ExportComponent} from './export/export.component';
+import { FiltersComponent } from './filters/filters.component';
 
 @NgModule({
     declarations: [
@@ -53,6 +58,7 @@ import {ExportComponent} from './export/export.component';
         ListCompaniesComponent,
         MapCompaniesComponent,
         ExportComponent,
+        FiltersComponent,
     ],
     imports: [
         BrowserModule,
@@ -92,9 +98,12 @@ import {ExportComponent} from './export/export.component';
         MatStepperModule,
         AppRoutingModule,
         HttpClientModule,
+        LoadingBarHttpClientModule,
+        LoadingBarModule.forRoot(),
         AgmCoreModule.forRoot({
-            apiKey: 'AIzaSyCUMXjVSSgh_WRBmIYFH2sWAYKc_8-Fwl4'
-        })
+            apiKey: myGlobals.ApiKey
+        }),
+        AgmSnazzyInfoWindowModule
     ],
     providers: [],
     bootstrap: [AppComponent]
