@@ -1,20 +1,40 @@
-import { Component, OnInit } from '@angular/core';
-
+import {Component, Injectable, OnInit} from '@angular/core';
+import {RetrieveCompaniesService} from '../retrieve-companies.service';
+import {Company} from '../Model/company';
+import {HttpClient} from '@angular/common/http';
 
 @Component({
-  selector: 'app-map-companies',
-  templateUrl: './map-companies.component.html',
-  styleUrls: ['./map-companies.component.scss']
+    selector: 'app-map-companies',
+    templateUrl: './map-companies.component.html',
+    styleUrls: ['./map-companies.component.scss']
 })
+@Injectable()
 export class MapCompaniesComponent implements OnInit {
 
-  title: 'test';
-  lat: number = 51.678418;
-  lng: number = 7.809007;
+    companies: Company[];
+    styles;
 
-  constructor() { }
+    title: 'Google Map';
+    lat: number;
+    lng: number;
 
-  ngOnInit() {
-  }
+    constructor(private retrieveCompaniesService: RetrieveCompaniesService, private http: HttpClient) {
+        this.retrieveCompaniesService.retrieveCompanies.subscribe((companies: Company[]) => {
+            this.companies = companies;
+        });
+    }
 
+    ngOnInit() {
+        this.retrieveCompaniesService.getCompanies();
+        if (navigator.geolocation) {
+            navigator.geolocation.getCurrentPosition((pos) => {
+                this.lat = pos.coords.latitude;
+                this.lng = pos.coords.longitude;
+            });
+        }
+        this.http.get('./assets/mapstyle.json').subscribe(data => {
+          this.styles = data['styles'];
+        });
+
+    }
 }
